Refactor InstructionHandler function signature
T (*)(std::unique_ptr<Processor> &processor) ->
T (*)(std::unique_ptr<Processor> &processor, uint8_t opcode)

The reason for this refactoring is that some instructions encode
the parameters in the opcode, like: BBRi, BBSi, RMBi and SMBi.
All instructions encode the addressing mode in the opcode, so in the
future we could use the same handler and use the new parameter to
use specific addressing modes too.